+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2025-01-19
+
 ### Added
 
 - Added `env-filter` feature to `tracing-subscriber`.
@@ -44,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump redis to v0.28.
 
-[unreleased]: https://github.com/4lkaid/axum-kit/compare/v0.5.2...HEAD
+[unreleased]: https://github.com/4lkaid/axum-kit/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/4lkaid/axum-kit/compare/v0.5.2...v0.6.0
 [0.5.2]: https://github.com/4lkaid/axum-kit/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/4lkaid/axum-kit/compare/v0.5.0...v0.5.1